project

+++ app/Traits/FileUploadTrait.php
@@ -8,22 +8,22 @@ use Illuminate\Support\Facades\Storage;
 trait FileUploadTrait
 {
  
-    public function storeFile($data){
+    public function storeFile($data, $folder){
         
         if(Arr::has($data, 'avatar')){
             $data['avatar'] = (Arr::pull($data, 'avatar'));
-            $data['avatar'] = (Arr::pull($data, 'avatar'))->store('user-avatar');
+            $data['avatar'] = (Arr::pull($data, 'avatar'))->store($folder);
         }
 
         if(Arr::has($data, 'nid')){
             $data['nid'] = (Arr::pull($data, 'nid'));
-            $data['nid'] = (Arr::pull($data, 'nid'))->store('user-nid');
+            $data['nid'] = (Arr::pull($data, 'nid'))->store($folder);
         }
 
         return $data;
     }
 
-    public function updateFile($object, $data){
+    public function updateFile($object, $data, $folder){
 
 
         if(Arr::has($data, 'avatar')){
@@ -33,7 +33,7 @@ trait FileUploadTrait
             }
 
             $data['avatar'] = (Arr::pull($data, 'avatar'));
-            $data['avatar'] = (Arr::pull($data, 'avatar'))->store('user-avatar');
+            $data['avatar'] = (Arr::pull($data, 'avatar'))->store($folder);
         }
 
         if(Arr::has($data, 'nid')){
@@ -43,7 +43,7 @@ trait FileUploadTrait
             }
 
             $data['nid'] = (Arr::pull($data, 'nid'));
-            $data['nid'] = (Arr::pull($data, 'nid'))->store('user-nid');
+            $data['nid'] = (Arr::pull($data, 'nid'))->store($folder);
         }
 
         return $data;
@@ -51,11 +51,17 @@ trait FileUploadTrait
     }
 
     public function destroyFile($object):void{
-        if (Storage::exists($object->avatar)) {
-            Storage::delete($object->avatar);
+        if(Arr::has($object, 'avatar')){
+            
+            if (Storage::exists($object->avatar)) {
+                Storage::delete($object->avatar);
+            }
         }
-        if (Storage::exists($object->nid)) {
-            Storage::delete($object->nid);
+        if(Arr::has($object, 'nid')){
+
+            if (Storage::exists($object->nid)) {
+                Storage::delete($object->nid);
+            }
         }
     }
 }
